+++ lib/twitter_api_v2.dart
@@ -23,6 +23,8 @@ export 'package:twitter_api_v2/src/service/spaces/space_expansion.dart';
 export 'package:twitter_api_v2/src/service/spaces/space_meta.dart';
 export 'package:twitter_api_v2/src/service/spaces/space_state.dart';
 export 'package:twitter_api_v2/src/service/spaces/spaces_service.dart';
+export 'package:twitter_api_v2/src/service/tweets/filtering_rule_data.dart';
+export 'package:twitter_api_v2/src/service/tweets/filtering_rule_meta.dart';
 export 'package:twitter_api_v2/src/service/tweets/referenced_tweet_data.dart';
 export 'package:twitter_api_v2/src/service/tweets/reply_setting.dart';
 export 'package:twitter_api_v2/src/service/tweets/tweet_attachments.dart';